bootstrap upgrade / jquery removal
