+++ src/modules/web/locales/fr.json
@@ -198,9 +198,6 @@
         },
         "tips": [
             "<div class='headline'>Inviter des amis</div> <div class='text'>Vous pouvez enregistrer encore un autre réseau wi-fi, pour chaque utilisateur qui entre à wifis.org par <a href='{0}' target='_blank' title='Voilà votre lien personnel pour inviter des amis à wifis.org'>votre lien d’invitation personnelle</a> </div> ",
-                
-            "<div class='headline'>URLs de réseaux wi-fi courtes</div> <div class='text'>Utilisez l’URL qui vous préférez: {1} </div>",
-            
             "<div class='headline'>Plus de conseils</div> <div class='text'> <b>N’oubliez pas de changer le nom de votre wi-fi</b><br> Après avoir enregistré un nom wi-fi, n’oubliez pas de aussi renommer votre réseau wifi par ce nom. En savoir plus sous <a href='/{__lang}/faq'>FAQs</a>. <br><br> <b>Un nom wifis.org pour plusieurs routeurs</b><br> Vous pouvez ajouter les caractères que vous voulez à votre nom en ajoutant une barre oblique derrière votre nom  wifis.org et ainsi utiliser le nom pour plusieurs routeurs: {2} <b>Majuscule/minuscule</b><br> Des noms de wi-fi peuvent contenir des majuscules. <a href='http://wifis.org/iphONE'>wifis.org/iphONE</a> devient <a href='http://wifis.org/iPhone'>wifis.org/iPhone</a> <br><br> <b>Aidez-nous</b><br> Soutenez ce service gratuit à l’aide de Flattr ou en parlant de notre service a vos amis.</div>"
         ]
     },
@@ -209,7 +206,11 @@
         "text": {
             "text": "Le propriétaire de ce wifi utilise un nom de wifs.org expressément, pour que VOUS puissiez lui écrire un message. Il veut vous donner la possibilité de pouvoir simplement lui dire bonjour ou de partager sa connexion d’internet.",
             "notfound": "Il n’y a pas de wi-fi enregistré sous ce nom. Vérifiez si vous avez écrit l’adresse correctement.</p> <p>Pour savoir plus sur  wifis.org, cliquez <a href='/'>ici</a>.",
-            "mywifi": "<h4>C'est votre Wi-Fi</h4> N'hésitez pas à essayer le formulaire de contact."
+            "mywifi": "<h4>C'est votre Wi-Fi</h4> N'hésitez pas à essayer le formulaire de contact.<br><br><b>Indice:</b> Utilisez l’URL qui vous préférez:"
+        },
+        "urls": {
+            "subdomains": "Sous-domaine",
+            "short": "URL courte"
         },
         "form": {
             "contact": "Votre adresse émail / numéro de téléphone <span style='font-weight: normal'>(Comment peut-on vous contacter?)</span>",
