+++ src/modules/web/locales/nl.json
@@ -195,9 +195,6 @@
         },
         "tips": [
             "<div class='headline'>ontvang een ander WiFi ID</div> <div class='text'>Ieder van je vrienden die wifis.org die toetreedt via <a href='{0}' target='_blank' title=' Dit is je persoonlijke link om vrienden uit te nodigen op wifis.org'>je persoonlijke uitnodigingslink</a>, zal de limiet van je wifis.org ID´s die je kan registreren met 1 verhogen.</div>",
-                
-            "<div class='headline'>verkorte url</div> <div class='text'>Kies hier je favoriet. Ze werken allemaal op dezelfde manier: {1} </div>",
-            
             "<div class='headline'>meer tips</div> <div class='text'> <b>Herbenoem je Wi-fi netwerk</b><br> Vergeet niet om je Wi-fi network te herbenoemen nadat je je wifis.org ID geregistreerd hebt. Kijk op <a href='/{__lang}/faq'>Veelgestelde vragen</a>. <br><br> <b>Eén ID voor meerdere routers</b><br> Om je wifis.org ID voor meerdere Wi-fi routers te gebruiken, kan je eender welk karakter toevoegen na een slash“/” achter jouw ID: {2} <b>Case sensitivity</b><br> wifis.org IDs zijn niet case sensitive. <a href='http://wifis.org/iphONE'>wifis.org/iphONE</a> wordt <a href='http://wifis.org/iPhone'>wifis.org/iPhone</a> <br><br> <b>Vertel het voort!</b><br> Steun onze gratis service door de Flattr knop links te gebruiken en door je vrienden te informeren. </div>"
         ]
     },
@@ -206,8 +203,13 @@
         "text": {
             "text": "De bezitter van dit Wi-fi netwerk registreerde zich bij wifis.org om je de kans te geven met hem in contact te komen. Je kan via deze weg shared internet toegang vragen of gewoon eens ´gedag´. In ieder geval, blijf steeds vriendelijk, je buur zal je dankbaar zijn.",
             "notfound": "Deze url is nog niet geregistreerd.Zorg ervoor dat je de correcte url gebruikt.</p> <p>Om deze url te registreren ga <a href='/p/login/'>hierheen</a>.",
-            "mywifi": "<h4>Dit is je Wifi.</h4> Probeer het contactformulier gerust eens uit."
+            "mywifi": "<h4>Dit is je Wifi.</h4> Probeer het contactformulier gerust eens uit.<br><br><b>Tip:</b> Kies hier je favoriet. Ze werken allemaal op dezelfde manier:"
         },
+        "urls": {
+            "standard": "Standaard",
+            "short": "Verkorte url"
+        },
+        
         "form": {
             "contact": "geef een e-mailadres of telefoonnummer in <span style='font-weight: normal'>(of iedere manier om met je in contact te komen)</span>",
             "message": "je boodschap",
