+++ src/modules/web/locales/ru.json
@@ -195,10 +195,7 @@
             "whileWifiDelete": "Опаньки. Ошибка при удалении."
         },
         "tips": [
-            "<div class='headline'>получить еще Wi-Fis</div> <div class='text'>Пригласив друга в wifis.org используя <a href='{0}' target='_blank' title='Это ваша персональная ссылка на wifis.org'>вашу персональную ссылку</a>, даст Вам возможность зарегистрировать еще один Wi-Fis.</div>",
-                
-            "<div class='headline'>короткие ссылки</div> <div class='text'>Выберите адрес, который больше нравиться. Они все работают одинаково: {1} </div>",
-            
+            "<div class='headline'>получить еще Wi-Fis</div> <div class='text'>Пригласив друга в wifis.org используя <a href='{0}' target='_blank' title='Это ваша персональная ссылка на wifis.org'>вашу персональную ссылку</a>, даст Вам возможность зарегистрировать еще один Wi-Fis.</div>",                
             "<div class='headline'>подсказки</div> <div class='text'> <b>Новое имя вашей Wi-Fi сети</b><br> Не забудьте переименовать вашу Wi-Fi сеть после регистрации Wi-Fis. Смотрите <a href='/{__lang}/faq'>ЧаВо</a>. <br><br> <b>Один Wi-Fis для нескольких маршрутизаторов</b><br> Для использования одного Wi-Fis для нескольких Wi-Fi маршрутизаторов, достаточно добавить любые символы после '/' за вашим Wi-Fis: {2} <b>Чувствительность к регистру</b><br> Wi-Fis НЕ чувствительны к регистру. <a href='http://wifis.org/iphONE'>wifis.org/iphONE</a> и <a href='http://wifis.org/iPhone'>wifis.org/iPhone</a> - это одна и та же ссылка <br><br> <b>Расскажите о нас!</b><br> Поддержите нас, нажав на кнопку Flattr слева, расскажите о нас друзьям. </div>"
         ]
     },
@@ -207,7 +204,11 @@
         "text": {
             "text": "Владелец этой Wi-Fi сети пользуется сервисом wifis.org, чтобы дать Вам шанс связаться с ним.  Вы, возможно, хотите воспользоваться его сетью или просто сказать 'Привет!'. В любом случае, будьте добры и Ваши соседи будут вам благодарны.",
             "notfound": "Такая Wi-Fi сеть еще не зарегистрирована. Убедитесь что вы перешли по правильному адресу.</p> <p>Если вы хотите чтобы эта страница была вашей, то перейдите <a href='/p/login/'>по этой ссылке</a>.",
-            "mywifi": "<h4> Это ваш Wi-Fi.</h4> Не стесняйтесь попробовать контактную форму."
+            "mywifi": "<h4>Это ваш Wi-Fi.</h4> Не стесняйтесь попробовать контактную форму.<br><br><b>Намекать:</b> Выберите адрес, который больше нравиться. Они все работают одинаково:"
+        },
+        "urls": {
+            "standard": "Стандарт",
+            "short": "короткие ссылки"
         },
         "form": {
             "contact": "пожалуйста, укажите ваш адрес электронной почты или номер телефона <span style='font-weight: normal'>(для связи с вами)</span>",
