+++ src/modules/web/views/home.hbs
@@ -27,8 +27,6 @@
     </div>
 </div>
 
-<div class="container-lg">
-    <div class="wifichecker">
-        <div class="mb-2 text-center fw-bold">{{{__ "home.wifichecker"}}} </div>
-        {{> wifiForm }}
-    </div>
+<div class="container-lg no-js-hidden">
+    <div class="mb-2 text-center fw-bold">{{{__ "home.wifichecker"}}} </div>
+    {{> wifiForm }}
