+++ src/modules/web/views/page.hbs
@@ -1,1 +1,1 @@
-<!-- -->
+{{!-- Everything is in the default layout.hbs, but doesn't work without a view file --}}
