+++ src/modules/web/views/partials/footer.hbs
@@ -132,20 +132,8 @@
     <div class="d-flex flex-column flex-sm-row justify-content-between py-3 border-top">
         <p>{{{__ "label.footer.text"}}}</p>
         <ul class="list-unstyled d-flex w-30">
-            {{!--
             <li class="ms-3">
-                <a class="link-dark icon twitter_icon" href="https://twitter.com/wifis_org">
-                    <span>twitter</span>
-                </a>
-            </li>
-            <li class="ms-3">
-                <a class="link-dark icon facebook_icon" href="https://www.facebook.com/wifis.org/">
-                    <span>facebook</span>
-                </a>
-            </li>
-            --}}
-            <li class="ms-3">
-                <a class="link-dark icon github_icon nav-link text-muted text-nowrap w-100"
+                <a class="link-dark icon github_icon nav-link text-muted text-nowrap w-100" rel='nofollow'
                     href="https://github.com/madmaxmatze/wifis.org">
                     <span>github</span>
                 </a>
