--- src/modules/web/views/partials/head_js.hbs
@@ -1,10 +1,0 @@
- <script>
-{{#if (equals env "production")}}
-(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-})(window,document,'script','dataLayer','GTM-M6XKF9N');
-{{/if}}
-document.documentElement.classList.replace("no-js", "js");
-</script>
