+++ src/modules/web/views/wifi.hbs
@@ -12,22 +12,21 @@
 
             <ul>
                 <li>
-                    <b>Standard:</b>
-                    <a href='https://wifis.org/{{wifiId}}'>wifis.org/{{wifiId}}</a>
+                    <b>{__wifi.urls.standard}:</b>
+                    <a href='https://wifis.org/{{wifiId}}' rel='nofollow'>wifis.org/{{wifiId}}</a>
                 </li>
                 <li>
                     <b>{__wifi.urls.subdomains}:</b>
-                    <a href='https://{{wifiId}}.wifis.org'>{{wifiId}}.wifis.org</a>
+                    <a href='https://{{wifiId}}.wifis.org' rel='nofollow'>{{wifiId}}.wifis.org</a>
                 </li>
                 <li>
                     <b>{__wifi.urls.short}:</b>
-                    <a href='https://wif.is/{{wifiId}}'>wif.is/{{wifiId}}</a>
+                    <a href='https://wif.is/{{wifiId}}' rel='nofollow'>wif.is/{{wifiId}}</a>
                 </li>
                 <li>
                     <b>{__wifi.urls.subwifis}:</b> 
-                    <a href='https://wifis.org/{{wifiId}}/whatever/sub/wifi_$!@'>wifis.org/{{wifiId}}/whatever/sub/wifi_$!@</a>
-                </li>
-                
+                    <a href='https://wifis.org/{{wifiId}}/whatever/sub/wifi_$!@' rel='nofollow'>wifis.org/{{wifiId}}/whatever/sub/wifi_$!@</a>
+                </li>        
             </ul>
         </div>
     {{/if}}
